# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Uno
  version: "1"
  description: |
    Uno's HTTP surface: the builder API under `/api`, which the SPA calls, and
    the public Stremio-protocol addon under `/u/{token}`, which Nuvio calls.
    Everything else on the server is the embedded SPA and is not described
    here.

    **Errors.** Every error under `/api`, an unknown path included, is JSON
    (`Error`): `error` holds one line of words and, for an error the builder
    acts on by kind, `code` names it (`ProfileNotFound`). Push is the one
    exception: past auth, profile resolution and body decoding it answers
    `PushResult` whether it succeeds or fails. The addon under `/u/{token}`
    answers errors in plain text (`text/plain; charset=utf-8`), which its
    clients don't read.

    **Request bodies.** Every body is decoded strictly: a field the request
    doesn't have, at any depth, is a 400 naming it (`invalid request body:
    unknown field "name"`). The default limit is 1 MiB, past which the answer
    is 413.

    **Auth.** Every `/api` operation except `GET /api/health` and
    `GET /api/config` takes a Nuvio access token as `Authorization: Bearer`.
    Any of them can answer 401 (missing, malformed, invalid or expired
    token), 403 (an account the server's allowlist doesn't list) or 502
    (`auth service unavailable`: Nuvio's signing keys couldn't be fetched).
    Any operation can answer 500 for a failure of Uno's own.

...

...
    TMDB key or TMDB refuses it.

...

...
servers:
...
security:
  - nuvioBearer: []
paths:
...
  /api/profiles:
    get:
      tags: [profiles]
      operationId: listProfiles
      summary: The caller's Nuvio profiles
      description: Nuvio's profile list, each with the picture Nuvio's apps show for it.
      responses:
        "200":
          description: Every profile on the caller's Nuvio account.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PickerProfile"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "502": { $ref: "#/components/responses/NuvioUnavailable" }
  /api/profiles/select:
    post:
      tags: [profiles]
      operationId: selectProfile
      summary: Pick a Nuvio profile to build for
      description: |
        Matches `profile_index` against the caller's live Nuvio profiles and
        resolves the Uno profile for that slot, creating it on first pick.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [profile_index]
              properties:
                profile_index:
                  type: integer
      responses:
        "200":
          description: The Uno profile, with its addon manifest URL.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SelectedProfile"
        "400":
          description: The body doesn't decode, or the caller's Nuvio account has no profile at that index.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "413": { $ref: "#/components/responses/PayloadTooLarge" }
        "502": { $ref: "#/components/responses/NuvioUnavailable" }
...
  /api/catalogs/preview:
    post:
      tags: [recipes]
      operationId: previewCatalog
      summary: Run a recipe against TMDB
      description: |
        Returns the first page of titles a recipe finds, saving nothing. The
        provider is always `tmdb`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [type, params]
              properties:
                type:
                  $ref: "#/components/schemas/CatalogType"
                params:
                  $ref: "#/components/schemas/RecipeParamsString"
                genre:
                  type: string
                  description: A genre name from the recipe's genre options, narrowing the preview.
      responses:
        "200":
          description: The first page of titles.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Preview"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "413": { $ref: "#/components/responses/PayloadTooLarge" }
        "422": { $ref: "#/components/responses/KeyProblem" }
        "502": { $ref: "#/components/responses/TMDBUnavailable" }
  /api/catalogs/genre-options:
    post:
      tags: [recipes]
      operationId: catalogGenreOptions
      summary: The genres a recipe can be narrowed by
      description: The same list the manifest advertises as the catalog's genre extra.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [type, params]
              properties:
                type:
                  $ref: "#/components/schemas/CatalogType"
                params:
                  $ref: "#/components/schemas/RecipeParamsString"
      responses:
        "200":
          description: The genre options.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Genre"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "413": { $ref: "#/components/responses/PayloadTooLarge" }
        "422": { $ref: "#/components/responses/KeyProblem" }
        "502": { $ref: "#/components/responses/TMDBUnavailable" }
...
components:
  securitySchemes:
    nuvioBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: A Nuvio access token, verified against Nuvio's signing keys.
...
  responses:
    BadRequest:
      description: The request is malformed or refused in words the caller can act on.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    Unauthorized:
      description: Missing, malformed, invalid or expired bearer token.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: The server's allowlist doesn't list this Nuvio account.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    PayloadTooLarge:
      description: The body is over its cap (1 MiB; 4 MiB for the import routes).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    KeyProblem:
      description: Per-account key server only. The account has no TMDB key yet, or TMDB doesn't accept it.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    TMDBUnavailable:
      description: TMDB couldn't be reached, or the auth service is unavailable.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NuvioUnavailable:
      description: A Nuvio call failed, or the auth service is unavailable.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
    CatalogType:
      type: string
      enum: [movie, series]
...
    RecipeParams:
      type: object
      description: |
        A TMDB recipe. Which fields apply depends on the catalog type; a field
        the type doesn't know is refused. Rules: `docs/data-model.md`,
        *Recipe params (TMDB)*.
      properties:
        randomized: { type: boolean }
        sort_by: { type: string }
        with_genres: { type: string, description: "TMDB genre ids, comma (AND) or pipe (OR) separated." }
        without_genres: { type: string }
        with_original_language: { type: string }
        vote_average_gte: { type: number }
        vote_average_lte: { type: number }
        vote_count_gte: { type: integer }
        vote_count_lte: { type: integer }
        with_runtime_gte: { type: integer }
        with_runtime_lte: { type: integer }
        with_watch_providers: { type: string, description: Requires watch_region. }
        watch_region: { type: string }
        with_companies: { type: string }
        without_companies: { type: string }
        with_keywords: { type: string }
        without_keywords: { type: string }
        certification: { type: string }
        certification_gte: { type: string }
        certification_lte: { type: string }
        certification_country: { type: string, description: Requires any certification field. }
        primary_release_date_gte: { type: string, format: date, description: movie only. }
        primary_release_date_lte: { type: string, format: date, description: movie only. }
        released_within_days: { type: integer, description: "movie only; excludes the primary_release_date fields." }
        with_collection: { type: string, description: "movie only; a TMDB collection id. Excludes every other field but randomized." }
        first_air_date_gte: { type: string, format: date, description: series only. }
        first_air_date_lte: { type: string, format: date, description: series only. }
        aired_within_days: { type: integer, description: "series only; excludes the first_air_date fields." }
        with_networks: { type: string, description: series only. }
    RecipeParamsString:
      type: string
      description: A `RecipeParams` object, JSON-encoded.
      contentMediaType: application/json
      contentSchema:
        $ref: "#/components/schemas/RecipeParams"
      example: '{"sort_by":"popularity.desc","with_genres":"28"}'
...
    PickerProfile:
      type: object
      description: One Nuvio profile as Nuvio lists it, plus the picture Nuvio's apps show.
      required: [id, user_id, profile_index, name, uses_primary_addons, avatar_color_hex, avatar_id, avatar_url, pin_enabled, avatar_image_url]
      properties:
        id: { type: string, description: Nuvio's profile id. }
        user_id: { type: string }
        profile_index: { type: integer }
        name: { type: string }
        uses_primary_addons:
          type: boolean
          description: The profile uses profile 1's addons; push refuses it.
        avatar_color_hex: { type: string }
        avatar_id: { type: string }
        avatar_url: { type: string }
        pin_enabled: { type: boolean }
        avatar_image_url:
          type: string
          description: Its own upload, else its built-in avatar's image; "" when it shows its colour.
    SelectedProfile:
      type: object
      required: [manifest_url]
      properties:
        manifest_url: { type: string, format: uri, description: The profile's addon manifest URL. }
...
    Preview:
      type: object
      required: [randomized, items, total_results]
      properties:
        randomized:
          type: boolean
          description: The recipe shuffles, so page 1 won't match what Nuvio shows.
        items:
          type: array
          items:
            $ref: "#/components/schemas/PreviewItem"
        total_results:
          type: integer
          description: TMDB's count of matches across every page.
    PreviewItem:
      type: object
      required: [tmdb_id, title, year, poster]
      properties:
        tmdb_id: { type: integer }
        title: { type: string }
        year: { type: string }
        poster: { type: string }
...
    Error:
      type: object
      description: The answer to every failed request under `/api` but push's own.
      required: [error]
      properties:
        error:
          type: string
          description: One line of words, for a person to read.
        code:
          type: string
          description: A stable name for an error the builder acts on by kind rather than by status; absent when there is none.
...
    Genre:
      type: object
      required: [id, name]
      properties:
        id: { type: integer }
        name: { type: string }
...